refactor(renderer): BET-415 extract useClickAway hook + SegmentedBar (duplication gate)
- useClickAway hook: shared click-outside + Escape dismiss for popovers/menus
  (was duplicated inline in SessionHeader ContextPill + SessionMenu + ModelPicker)
- SegmentedBar component: shared fresh/cache-write/cache-read segment rendering
  (was duplicated in the mini pill bar + the popover bar)
- Clears all SessionHeader self-clones; remaining gate clones are test setup
  boilerplate in the harness file (intra-file, advisory only)
